Savestates: Asynchronous compression
