+++ rpcs3/Crypto/unzip.h
@@ -16,7 +16,7 @@ inline bool unzip(const std::vector<u8>& src, fs::file& out)
 	return unzip(src.data(), src.size(), out);
 }
 
-bool zip(const void* src, usz size, fs::file& out);
+bool zip(const void* src, usz size, fs::file& out, bool multi_thread_it = false);
 
 template <typename T>
 inline bool zip(const T& src, fs::file& out)
