+++ rpcs3/Emu/CMakeLists.txt
@@ -50,6 +50,7 @@ target_sources(rpcs3_emu PRIVATE
     ../util/dyn_lib.cpp
     ../util/sysinfo.cpp
     ../util/cpu_stats.cpp
+    ../util/serialization_ext.cpp
     ../../Utilities/bin_patch.cpp
     ../../Utilities/cheat_info.cpp
     ../../Utilities/cond.cpp
