+++ rpcs3/Loader/TAR.cpp
@@ -9,7 +9,7 @@
 
 #include "util/asm.hpp"
 
-#include "Emu/savestate_utils.hpp"
+#include "util/serialization_ext.hpp"
 
 #include <charconv>
 #include <span>
@@ -139,7 +139,7 @@ std::unique_ptr<utils::serial> tar_object::get_file(const std::string& path, std
 		}
 		else
 		{
-			tar_log.error("tar_object::get_file() failed to parse header: offset=0x%x, filesize=0x%x", largest_offset, max_size);
+			tar_log.error("tar_object::get_file() failed to parse header: offset=0x%x, filesize=0x%x", offset, max_size);
 		}
 
 		return { size, {} };
